refactor: Password 제거, profileUrl 추가에 따른 테스트 코드 수정
